requirement.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.0
django-shortuuidfield==0.1.3
get==2019.4.13
idna==2.8
mysqlclient==1.4.4
Pillow==6.2.1
post==2019.4.13
public==2019.4.13
python-memcached==1.59
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
urllib3==1.25.7
